Publish missing initial/IReady states for modules with unfulfilled interface state contracts
Several modules implement interfaces that formally require a state (via
interface inheritance, e.g. IFilters->IMotion->IReady, IAutoGuiding->IExposureTime)
but never publish it, leaving any subscriber to retry and warn indefinitely:

- AutoGuiding/ScienceFrameAutoGuiding never published IExposureTime unless
  set_exposure_time() was called externally.
- PipelineCamera never published any state, including the IExposure required
  by ICamera.
- FlatField never published IMotion, required via IFilters.
- MotionStatusMixin only published IMotion, never the separately-enumerated
  IReady that IMotion implies, leaving telescopes (and anything else using
  the mixin without its own IReady handling) with no ready state at all.
  BaseRoof's duplicate ready-derivation and DummyMode's hardcoded publish
  are now folded into the mixin's shared default.
